Truncate logs if they exceed 13500 characters
Once the message card exceeds a certain amount of characters, the teams webhook is
unable to curl it without erroring complaining of large payload or with
a 413 error.
We are truncating the error log at 13500 characters (max we could found
we could send) to allow some wiggle room for the other message cards fields,
but are supplying a way to override that value to be included in the pipeline under
the resource type source parameters parameters.
Readme and test pipeline updated to reflect new variable added.