If a key in the logs file contains a slash then the taskmap get confused
We noticed a keyerror can occur when a key in the logs dict contains a
slash.
Hence we remove the slash to allow the processing of the logs dict into
the Teams output message.
The key with slash only seems to occur if the logs contain a get image
log. If they do not it will happiky run through. This fix should
workaround that issue though.